Share usd-formatting utility function across files
