+++ test/monthly_sales_alt_test.py
@@ -3,20 +3,7 @@
 import os
 import pandas
 
-from app.monthly_sales_alt import to_usd, top_selling_products
-
-def test_to_usd():
-    # it should apply USD formatting
-    assert to_usd(4.50) == "$4.50"
-
-    # it should display two decimal places
-    assert to_usd(4.5) == "$4.50"
-
-    # it should round to two places
-    assert to_usd(4.55555) == "$4.56"
-
-    # it should display thousands separators
-    assert to_usd(1234567890.5555555) == "$1,234,567,890.56"
+from app.monthly_sales_alt import top_selling_products
 
 def test_top_selling_products():
     # it should return a list of names and sales totals, sorted by sales ascending with top sellers first:
